+++ R/scSeurat.R
@@ -2,7 +2,7 @@
 #'
 #' @param path_10x Path to 10X RNAseq data "filtered_feature_bc_matrix" folder
 #' @param h5_file Path to 10X h5 file
-#' @param frag_file Path to 10X ATAC fragments file
+#' @param frag_file Path to 10X ATAC fragments file. Only required for ATAC data
 #' @param min_cells Passed to CreateSeuratObject: Include features detected in
 #'     at least this many cells. Will subset the counts matrix as well. To
 #'     reintroduce excluded features, create a new object with a lower cutoff.
@@ -81,11 +81,27 @@ tenx_load_qc <- function(path_10x = "",
                             species_pattern,
                             remove_species_pattern)
 
+        if (nrow(rna_raw_data) == 0) {
+            stop("No genes left in object after filtering using ",
+                 "species_pattern! Check species_pattern argument. ",
+                 "Before filtering gex data using species_pattern the first ",
+                 "genes were: ",
+                 paste(gex_first_ten_genes, collapse = "\n"))
+        }
+
         seurat <-
             Seurat::CreateSeuratObject(rna_raw_data,
                                        min.cells = min_cells,
-                                       min.features = min_features) %>%
-            Seurat::PercentageFeatureSet(pattern = gsub("_", "-", mt_pattern),
+                                       min.features = min_features)
+
+        if (ncol(seurat) == 0) {
+            stop("No data left after applying min.cells and min.features.",
+                 " Check your data and arguments.")
+        }
+
+        seurat <-
+            Seurat::PercentageFeatureSet(seurat,
+                                         pattern = gsub("_", "-", mt_pattern),
                                          col.name = "percent.mt")
             # Need to change all underscores to dashes due to CreateSeuratObject
             # doing the same
@@ -104,10 +120,21 @@ tenx_load_qc <- function(path_10x = "",
 
     if (grepl("ATAC", exp_type)) {
         # subset the data to only include the species of interest
+        atac_first_ten_peaks <- head(rownames(rna_raw_data), 10)
+
         atac_raw_data <-
             filter_raw_data(atac_raw_data,
                             species_pattern,
                             remove_species_pattern)
+
+        if (nrow(atac_raw_data) == 0) {
+            stop("No peaks left in object after filtering using ",
+                 "species_pattern! Check species_pattern argument. ",
+                 "Before filtering gex data using species_pattern the first ",
+                 "peaks were: ",
+                 paste(atac_first_ten_peaks, collapse = "\n"))
+        }
+
 
         if (exp_type == "ATAC") {
             seurat <-
@@ -134,13 +161,7 @@ tenx_load_qc <- function(path_10x = "",
 
     }
 
-    if (nrow(seurat) == 0) {
-        stop("!No data left in object! Check your species_pattern argument. ",
-              "Before filtering based on species_pattern there were ",
-              gex_orig_cells,
-              " cells and the first genes were ",
-              gex_first_ten_genes)
-    }
+
 
     if (violin_plot && grepl("GEX", exp_type)) {
         print(Seurat::VlnPlot(seurat,
@@ -286,7 +307,11 @@ gen_cellecta_bc_data <- function(file, verbose = FALSE, output = tempfile(),
 
   system(system_cmd)
 
-  results <- readr::read_delim(output, delim = "\t", col_names = TRUE)
+  results <-
+    readr::read_delim(output,
+                      delim = "\t",
+                      col_names = TRUE,
+                      show_col_types = FALSE)
 
   return(results)
 }
@@ -343,7 +368,10 @@ process_ltbc <- function(sobject, cid_lt, histogram = FALSE,
 
   # Integrate the lineage tracing barcode into the Seurat object metadata
   sobject$lt <-
-    cid_lt[stringr::str_remove(sobject@assays$RNA@counts@Dimnames[[2]], "-1$")]
+    cid_lt[stringr::str_remove(Seurat::Cells(sobject),
+                               "-1$")] %>%
+    as.vector()
+
 
   # Generate the frequency tables
   ylabel <- "Number of Cells"
@@ -606,6 +634,7 @@ process_seurat <- function(sobject,
 #' @param sd_up Number of standard deviations above the median to subset
 #' @param make_plots Whether to make plots of the features before and after subsetting
 #' @param features Vector of features to use for subsetting
+#' @param sample_name Name of sample to use in plot titles
 #'
 #' @return A Seurat object
 #' @export
@@ -619,7 +648,8 @@ auto_subset <- function(sobject,
                         sd_up = 2,
                         make_plots = TRUE,
                         features = c("nCount_RNA",
-                                     "nFeature_RNA")) {
+                                     "nFeature_RNA"),
+                        sample_name = NULL) {
 
     cutoff_table <-
         sobject@meta.data %>%
@@ -651,10 +681,10 @@ auto_subset <- function(sobject,
                                 values <= cutoffs$max_val[1])]
     }
 
-    if (make_plots) {
-        print(feature_hist(sobject,
-                           features = features,
-                           cutoff_table = cutoff_table))
-    }
+    # if (make_plots) {
+    #     print(feature_hist(sobject,
+    #                        features = features,
+    #                        cutoff_table = cutoff_table))
+    # }
     return(sobject)
 }